Put fitter modules in pahfit.fitters. Rename apfitter to ap_fitter.
