+++ pahfit/fitters/fitter.py
